# ⚡️ Speed up method `GoogleDriveIndexer.count_files_recursively` by 29%