<!-- CODEFLASH_OPTIMIZATION: {"function":"GoogleDriveIndexer.count_files_recursively","file":"unstructured_ingest/processes/connectors/google_drive.py","speedup_pct":"29%","speedup_x":"0.29x","original_runtime":"2.25 milliseconds","best_runtime":"1.74 milliseconds","optimization_type":"general","timestamp":"2025-08-22T10:35:37.636Z","version":"1.0"} -->
### 📄 29% (0.29x) speedup for ***`GoogleDriveIndexer.count_files_recursively` in `unstructured_ingest/processes/connectors/google_drive.py`***

⏱️ Runtime :   **`2.25 milliseconds`**  **→** **`1.74 milliseconds`** (best of `54` runs)
### 📝 Explanation and details


The optimization achieves a **29% speedup** by eliminating redundant computation in the extension filtering logic. 

**What was optimized:**
- **Pre-computed extension set**: Instead of creating `[e.lower() for e in extensions]` for every single file processed (line that took 13.4% of runtime), the extensions are converted to a lowercase `set` once at the beginning
- **O(1) set lookup**: Changed from O(n) list membership check to O(1) set membership check using `if file_ext in valid_exts`

**Why this is faster:**
In the original code, for each file encountered, a new list of lowercased extensions was created and then searched linearly. With 2,028 hits on that line consuming 13.4% of total runtime, this was a major bottleneck. The optimized version does this computation just once (46 hits, 0.5% of runtime) and uses set membership which is significantly faster for lookups.

**Performance characteristics:**
- **Small datasets**: Minimal impact (tests show 0-8% variance, mostly slower due to set creation overhead)  
- **Large datasets with extension filtering**: Dramatic improvements - the `test_large_folder_with_extension_filter` shows **112% speedup** when filtering 999 files, and `test_large_with_extension_filter` shows **110% speedup** for 1000 files
- **No extension filtering**: Negligible impact since the optimization only affects the filtering path

The optimization is most beneficial for scenarios with many files and extension filtering enabled, which matches typical Google Drive indexing use cases.



✅ **Correctness verification report:**

| Test                        | Status            |
| --------------------------- | ----------------- |
| ⚙️ Existing Unit Tests | 🔘 **None Found** |
| 🌀 Generated Regression Tests | ✅ **43 Passed** |
| ⏪ Replay Tests | 🔘 **None Found** |
| 🔎 Concolic Coverage Tests | 🔘 **None Found** |
|📊 Tests Coverage       | 100.0% |
<details>
<summary>🌀 Generated Regression Tests and Runtime</summary>

```python
from dataclasses import dataclass, field
from types import SimpleNamespace

# imports
import pytest  # used for our unit tests
# function to test
from googleapiclient.discovery import Resource as GoogleAPIResource
from unstructured_ingest.interfaces import Indexer
from unstructured_ingest.processes.connectors.google_drive import \
    GoogleDriveIndexer

# ------------- Helper Classes for Mocking Google Drive API -------------

class MockListResponse:
    """Mock the .execute() response of the files().list() call."""
    def __init__(self, files=None, nextPageToken=None):
        self._files = files if files is not None else []
        self._nextPageToken = nextPageToken

    def execute(self):
        return {
            "files": self._files,
            "nextPageToken": self._nextPageToken,
        }

class MockFilesClient:
    """
    Mocks the files().list() method of Google Drive API.
    Takes a mapping: folder_id -> list of file/folder dicts.
    Supports pagination via 'pageSize' and 'pageToken'.
    """
    def __init__(self, folder_structure):
        # folder_structure: dict of folder_id -> list of file/folder dicts
        self.folder_structure = folder_structure
        self.calls = []  # for debugging

    def list(self, supportsAllDrives, includeItemsFromAllDrives, spaces, q, fields, pageToken=None, pageSize=1000):
        # Extract folder_id from query string
        # query is like "'folder_id' in parents"
        folder_id = q.split("'")[1]
        files = self.folder_structure.get(folder_id, [])
        # Pagination
        if pageToken is None:
            start = 0
        else:
            start = int(pageToken)
        end = start + pageSize
        page_files = files[start:end]
        if end < len(files):
            nextPageToken = str(end)
        else:
            nextPageToken = None
        self.calls.append((folder_id, start, end))
        return MockListResponse(files=page_files, nextPageToken=nextPageToken)

# ------------- Unit Tests -------------

# ----------- BASIC TEST CASES -----------

def test_empty_folder():
    """Folder contains no files or folders."""
    structure = {"root": []}
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 9.21μs -> 9.50μs (3.11% slower)

def test_folder_with_only_files():
    """Folder contains only files, no subfolders."""
    structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain", "fileExtension": "txt"},
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "pdf"},
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 6.61μs -> 6.90μs (4.26% slower)

def test_folder_with_files_and_subfolders():
    """Folder contains files and one subfolder."""
    structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain", "fileExtension": "txt"},
            {"id": "sub1", "mimeType": "application/vnd.google-apps.folder"},
        ],
        "sub1": [
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "pdf"},
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 9.38μs -> 9.78μs (4.13% slower)

def test_nested_folders_multiple_levels():
    """Folder tree with multiple levels of nesting."""
    structure = {
        "root": [
            {"id": "sub1", "mimeType": "application/vnd.google-apps.folder"},
            {"id": "f1", "mimeType": "text/plain", "fileExtension": "txt"},
        ],
        "sub1": [
            {"id": "sub2", "mimeType": "application/vnd.google-apps.folder"},
        ],
        "sub2": [
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "pdf"},
            {"id": "f3", "mimeType": "image/png", "fileExtension": "png"},
        ],
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 11.6μs -> 11.5μs (0.854% faster)

def test_extension_filter_basic():
    """Count only files with a given extension, case-insensitive."""
    structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain", "fileExtension": "TXT"},
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "pdf"},
            {"id": "f3", "mimeType": "image/png", "fileExtension": "PNG"},
        ]
    }
    client = MockFilesClient(structure)
    # Only count .txt and .png files
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root", extensions=["txt", "png"]) # 9.54μs -> 10.2μs (6.04% slower)

def test_extension_filter_with_none_and_empty():
    """Files with missing or empty fileExtension should not be counted."""
    structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain", "fileExtension": None},
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": ""},
            {"id": "f3", "mimeType": "image/png", "fileExtension": "png"},
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root", extensions=["pdf", "png"]) # 8.69μs -> 8.34μs (4.15% faster)

# ----------- EDGE TEST CASES -----------

def test_folder_with_only_subfolders():
    """Folder contains only subfolders, no files."""
    structure = {
        "root": [
            {"id": "sub1", "mimeType": "application/vnd.google-apps.folder"},
            {"id": "sub2", "mimeType": "application/vnd.google-apps.folder"},
        ],
        "sub1": [],
        "sub2": [],
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 10.9μs -> 11.0μs (0.891% slower)

def test_deeply_nested_empty_folders():
    """Deep folder structure, but no files at any level."""
    structure = {
        "root": [{"id": "sub1", "mimeType": "application/vnd.google-apps.folder"}],
        "sub1": [{"id": "sub2", "mimeType": "application/vnd.google-apps.folder"}],
        "sub2": [{"id": "sub3", "mimeType": "application/vnd.google-apps.folder"}],
        "sub3": [],
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 12.1μs -> 12.4μs (2.89% slower)

def test_files_with_missing_extension_key():
    """Some files do not have a fileExtension key at all."""
    structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain"},  # no fileExtension
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "pdf"},
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root", extensions=["pdf"]) # 7.70μs -> 8.11μs (5.03% slower)



def test_files_with_mixed_case_extensions():
    """Extensions filter should be case-insensitive."""
    structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain", "fileExtension": "TxT"},
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "PDF"},
            {"id": "f3", "mimeType": "image/png", "fileExtension": "PnG"},
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root", extensions=["txt", "pdf"]) # 12.9μs -> 14.0μs (7.90% slower)

def test_files_with_duplicate_ids():
    """Files with duplicate IDs in different folders should all be counted."""
    structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain", "fileExtension": "txt"},
            {"id": "sub1", "mimeType": "application/vnd.google-apps.folder"},
        ],
        "sub1": [
            {"id": "f1", "mimeType": "application/pdf", "fileExtension": "pdf"},  # same id as above
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 9.81μs -> 9.84μs (0.315% slower)

def test_files_with_no_mimeType():
    """Files with missing mimeType should be counted as files (not folders)."""
    structure = {
        "root": [
            {"id": "f1", "fileExtension": "txt"},  # no mimeType
            {"id": "sub1", "mimeType": "application/vnd.google-apps.folder"},
        ],
        "sub1": [
            {"id": "f2", "fileExtension": "pdf"},  # no mimeType
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 9.43μs -> 9.22μs (2.19% faster)

def test_files_with_folder_mimeType_but_extension():
    """A folder with a fileExtension should still be treated as a folder."""
    structure = {
        "root": [
            {"id": "sub1", "mimeType": "application/vnd.google-apps.folder", "fileExtension": "pdf"},
            {"id": "f1", "mimeType": "application/pdf", "fileExtension": "pdf"},
        ],
        "sub1": [
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "pdf"},
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root", extensions=["pdf"]) # 11.1μs -> 11.5μs (3.71% slower)

# ----------- LARGE SCALE TEST CASES -----------

def test_large_flat_folder():
    """A flat folder with 999 files, no subfolders."""
    structure = {
        "root": [
            {"id": f"f{i}", "mimeType": "application/pdf", "fileExtension": "pdf"} for i in range(999)
        ]
    }
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 116μs -> 109μs (6.39% faster)

def test_large_nested_structure():
    """A folder with 10 subfolders, each with 99 files."""
    structure = {
        "root": [
            {"id": f"sub{i}", "mimeType": "application/vnd.google-apps.folder"} for i in range(10)
        ]
    }
    for i in range(10):
        structure[f"sub{i}"] = [
            {"id": f"f{i}_{j}", "mimeType": "application/pdf", "fileExtension": "pdf"} for j in range(99)
        ]
    client = MockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 136μs -> 137μs (0.298% slower)

def test_large_tree_with_mixed_files_and_folders():
    """A tree with 5 top-level subfolders, each with 5 subfolders, each with 20 files."""
    structure = {"root": []}
    for i in range(5):
        sub_id = f"sub{i}"
        structure["root"].append({"id": sub_id, "mimeType": "application/vnd.google-apps.folder"})
        structure[sub_id] = []
        for j in range(5):
            subsub_id = f"sub{i}_{j}"
            structure[sub_id].append({"id": subsub_id, "mimeType": "application/vnd.google-apps.folder"})
            structure[subsub_id] = [
                {"id": f"f{i}_{j}_{k}", "mimeType": "application/pdf", "fileExtension": "pdf"} for k in range(20)
            ]
    client = MockFilesClient(structure)
    # 5*5*20 = 500 files
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 120μs -> 120μs (0.806% slower)

def test_large_pagination():
    """Test that pagination is handled correctly (folder with 500 files, pageSize=100)."""
    # We'll tweak the MockFilesClient to use pageSize=100 in the test
    structure = {
        "root": [
            {"id": f"f{i}", "mimeType": "application/pdf", "fileExtension": "pdf"} for i in range(500)
        ]
    }
    class PaginationMockFilesClient(MockFilesClient):
        def list(self, *args, **kwargs):
            # Force pageSize=100 for this test
            kwargs["pageSize"] = 100
            return super().list(*args, **kwargs)
    client = PaginationMockFilesClient(structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 75.0μs -> 76.6μs (2.01% slower)

def test_large_with_extension_filter():
    """Large folder with mixed extensions, filter for a specific extension."""
    structure = {
        "root": [
            {"id": f"f{i}", "mimeType": "application/pdf", "fileExtension": "pdf"} if i % 2 == 0
            else {"id": f"f{i}", "mimeType": "text/plain", "fileExtension": "txt"}
            for i in range(1000)
        ]
    }
    client = MockFilesClient(structure)
    # Only even-indexed files are pdf (500 files)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root", extensions=["pdf"]) # 493μs -> 234μs (110% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
#------------------------------------------------
from dataclasses import dataclass, field

# imports
import pytest  # used for our unit tests
# function to test
from googleapiclient.discovery import Resource as GoogleAPIResource
from unstructured_ingest.interfaces import Indexer
from unstructured_ingest.processes.connectors.google_drive import \
    GoogleDriveIndexer

# Helper classes for mocking Google Drive API

class FakeListResponse:
    """Mock response for files().list().execute()"""
    def __init__(self, files, nextPageToken=None):
        self._files = files
        self._nextPageToken = nextPageToken

    def execute(self):
        return {"files": self._files, "nextPageToken": self._nextPageToken}


class FakeFilesClient:
    """
    Mocks the files().list().execute() API.
    The structure is:
        folder_id -> list of file dicts (with id, mimeType, fileExtension)
        Optionally, for pagination, a folder_id can map to a list of (files, nextPageToken) tuples.
    """
    def __init__(self, folder_structure, paginated=False):
        self.folder_structure = folder_structure
        self.paginated = paginated
        self.calls = []  # Track calls for debugging

    def list(self, supportsAllDrives, includeItemsFromAllDrives, spaces, q, fields, pageToken, pageSize):
        folder_id = q.split("'")[1]
        self.calls.append((folder_id, pageToken))
        if not self.paginated:
            files = self.folder_structure.get(folder_id, [])
            return FakeListResponse(files)
        else:
            # folder_structure[folder_id] is a list of (files, nextPageToken) tuples
            pages = self.folder_structure.get(folder_id, [])
            if not pageToken:
                # Return first page
                files, nextPageToken = pages[0]
                return FakeListResponse(files, nextPageToken)
            else:
                # Find the page with the given token
                for files, token in pages:
                    if token == pageToken:
                        # Return next page (or empty if last)
                        idx = pages.index((files, token))
                        if idx + 1 < len(pages):
                            files2, nextPageToken2 = pages[idx + 1]
                            return FakeListResponse(files2, nextPageToken2)
                        else:
                            return FakeListResponse([], None)
                return FakeListResponse([], None)

# -------------------- UNIT TESTS --------------------

# BASIC TEST CASES

def test_empty_folder():
    # Test counting files in an empty folder (should return 0)
    folder_structure = {
        "root": []
    }
    client = FakeFilesClient(folder_structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 7.03μs -> 7.44μs (5.52% slower)

def test_folder_with_files_only():
    # Test folder containing only files (no subfolders)
    folder_structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain", "fileExtension": "txt"},
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "pdf"},
        ]
    }
    client = FakeFilesClient(folder_structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 4.97μs -> 5.33μs (6.78% slower)

def test_folder_with_subfolders_and_files():
    # Test folder containing files and one subfolder with files
    folder_structure = {
        "root": [
            {"id": "f1", "mimeType": "text/plain", "fileExtension": "txt"},
            {"id": "sub1", "mimeType": "application/vnd.google-apps.folder"},
        ],
        "sub1": [
            {"id": "f2", "mimeType": "application/pdf", "fileExtension": "pdf"},
            {"id": "f3", "mimeType": "image/png", "fileExtension": "png"},
        ]
    }
    client = FakeFilesClient(folder_structure)
    codeflash_output = GoogleDriveIndexer.count_files_recursively(client, "root") # 7.82μs -> 8.03μs (2.62% slower)

def test_folder_with_multiple_levels_of_subfolders():
    # Test nested subfolders (root -> sub1 -> sub2)
    folder_structure = {
        "root": [
            {"id": "sub1", "mimeType": "application/vnd.google-apps.folder"},
        ],
        "sub1": [
            {"id": "sub2", "mimeType": "application/vnd.google-apps.folder"},
…